pyproject.toml
[project]
name = "fastapi-pgvector"
version = "0.1.0"
description = "FastAPI with pgvector integration"
requires-python = ">=3.11"
dependencies = [
    "fastapi",
    "uvicorn",
    "sqlalchemy",
    "alembic",
    "psycopg2-binary",
    "pgvector",
    "openai>=1.107.3",
    "python-dotenv>=1.1.1",
    "PyPDF2>=3.0.1",
    "python-docx>=1.2.0",
    "pydantic-settings>=2.10.1",
]

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
# Use uv as the installer
package-type = "application"
install.use-uv = true

[tool.pdm.scripts]
start = { shell = "cd server && uvicorn main:app --reload" }
db-init = { shell = "cd server && python init_db.py" }
migrate = { shell = "cd server && alembic upgrade head" }
makemigrations = { shell = "cd server && alembic revision --autogenerate -m 'auto'" }
test = { shell = "cd server && pytest" }

# Optional shorter aliases
s = "pdm run start"
mm = "pdm run makemigrations"
m = "pdm run migrate"

[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.4.2",
]

# Environment variables for scripts
[tool.pdm.scripts.env]
DATABASE_URL = "postgresql+psycopg2://postgres@localhost:5432/postgres"